delete old commands
