--- src/actions/index.js
@@ -1,15 +1,0 @@
-import { autoSolve } from './autoSolve.js';
-import { autoSubmit } from './autoSubmit.js';
-import { solve } from './solve.js';
-import { submit } from './submit.js';
-
-/**
- * Barrel for for "top level" actions which are executed by cli commands.
- */
-
-export {
-  autoSolve,
-  autoSubmit,
-  solve,
-  submit,
-};
