+++ src/cli/submit.js
@@ -1,5 +1,7 @@
-import { createChain } from './actionChain.js';
-import * as links from './links/index.js';
+import { Command } from 'commander';
+import { createChain } from '../actions/actionChain.js';
+import * as links from '../actions/links/index.js';
+import { dayArgument, partArgument } from './arguments.js';
 
 /**
  * The links which together make up the submit action.
@@ -31,4 +33,14 @@ const actionChain = createChain(submitLinks);
  * @param {Number} day
  * @param {Number} part
  */
-export const submit = async (day, part) => actionChain({ day, part });
+const submit = async (day, part) => actionChain({ day, part });
+
+/**
+ * Command which allows the user to submit a puzzles answer to advent of code.
+ */
+export const submitCommand = new Command()
+  .name('submit')
+  .description('Run the solution for the puzzle, then submit the answer to advent of code.')
+  .addArgument(dayArgument)
+  .addArgument(partArgument)
+  .action(submit);
